requirements.txt
# Core
python>=3.9
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
jupyter>=1.0.0

# Data Processing
opencv-python>=4.5.0
rasterio>=1.2.0
geopandas>=0.9.0
xarray>=0.18.0
dask>=2021.6.0

# Machine Learning
torch>=1.9.0
transformers>=4.11.0
tensorflow>=2.6.0
scikit-learn>=0.24.0
xgboost>=1.4.0

# API & Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Geospatial
folium>=0.12.0
shapely>=1.7.1
sentinelhub>=3.4.0
eartiles>=0.10.2

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
wandb>=0.12.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0
